Moduł 5.8 Role Użytkowników

+++ WebAPI/Controllers/V1/PostsController.cs
@@ -6,10 +6,14 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.OData.Query;
 using Microsoft.AspNetCore.Mvc;
 using Swashbuckle.AspNetCore.Annotations;
+using System.Security.Claims;
 using WebAPI.Filters;
-using WebAPI.helpers;
+using WebAPI.Helpers;
 using WebAPI.Wrappers;
 using Microsoft.AspNetCore.Authorization;
+using Infrastructure.Identity;
+using System.Security.Claims;
+
 
 namespace WebAPI.Controllers.V1
 {
@@ -30,7 +34,7 @@ namespace WebAPI.Controllers.V1
         [SwaggerOperation(Summary = "Retrieves sort fields")]
         [HttpGet("[action]")]
         public IActionResult GetSortFields()
-        {            
+        {
             return Ok(SortingHelper.GetSortFields().Select(x => x.Key));
         }
 
@@ -50,6 +54,7 @@ namespace WebAPI.Controllers.V1
         }
 
         [SwaggerOperation(Summary = "Retrieves all posts")]
+        [Authorize(Roles = UserRoles.Admin)]
         [HttpGet("[action]")]
         [EnableQuery]
         public IQueryable<PostDto> GetAll()
@@ -71,27 +76,44 @@ namespace WebAPI.Controllers.V1
         }
 
         [SwaggerOperation(Summary = "Create a new post")]
+        [Authorize(Roles = UserRoles.User)]
         [HttpPost]
         public async Task<IActionResult> Create(CreatePostDto newPost)
         {
-            var post = await _postService.AddNewPostAsync(newPost);
+            var post = await _postService.AddNewPostAsync(newPost, User.FindFirstValue(ClaimTypes.NameIdentifier));
             return Created($"api/posts/{post.Id}", new Response<PostDto>(post));
         }
 
-        [SwaggerOperation(Summary = "Update a existing post")]
+        [SwaggerOperation(Summary = "Update a exsisting post")]
+        [Authorize(Roles = UserRoles.User)]
         [HttpPut]
         public async Task<IActionResult> Update(UpdatePostDto updatePost)
         {
+            var userOwnsPost = await _postService.UserOwnsPostAsync(updatePost.Id, User.FindFirstValue(ClaimTypes.NameIdentifier));
+            if (!userOwnsPost)
+            {
+                return BadRequest(new Response<bool>() { Succeeded = false, Message = "You do not own this post" });
+            }
+
             await _postService.UpdatePostAsync(updatePost);
             return NoContent();
         }
 
         [SwaggerOperation(Summary = "Delete a specific post")]
-        [HttpDelete]
+        [Authorize(Roles = UserRoles.AdminOrUser)]
+        [HttpDelete("{id}")]
         public async Task<IActionResult> Delete(int id)
         {
+            var userOwnsPost = await _postService.UserOwnsPostAsync(id, User.FindFirstValue(ClaimTypes.NameIdentifier));
+            var isAdmin = User.FindFirstValue(ClaimTypes.Role).Contains(UserRoles.Admin);
+
+            if (!isAdmin && !userOwnsPost)
+            {
+                return BadRequest(new Response<bool>() { Succeeded = false, Message = "You do not own this post" });
+            }
+
             await _postService.DeletePostAsync(id);
             return NoContent();
-        }      
+        }
     }
 }
